Set up functionality for user to buy stock, and that his data is updated in the DB.

+++ TradingApplication/TradingApplication.Domain/Models/Asset.cs
@@ -4,6 +4,6 @@
    {
       public string Symbol { get; set; }
 
-      public decimal PricePerShare { get; set; }
+      public double PricePerShare { get; set; }
    }
 }
